// src/generated/agent.ts
export interface paths {
    "/{san}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get Agent by SAN
         * @description Get Agent demographic details
         */
        get: {
            parameters: {
                query?: never;
                header: {
                    /** @description API Key */
                    "x-humana-api-key": components["parameters"]["api-key"];
                };
                path: {
                    san: components["parameters"]["san"];
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Successful response */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["Agent"];
                    };
                };
                /** @description Error Response */
                default: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["Errors"];
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/{san}/plans": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get AOA Plans
         * @description Get list of plans an Agent is allowed to sell for AOA
         */
        get: {
            parameters: {
                query: {
                    fips: string;
                    state: string;
                    /** @description ISO Format Date, defaults to today's date. */
                    as_of_date: string;
                };
                header: {
                    /** @description API Key */
                    "x-humana-api-key": components["parameters"]["api-key"];
                };
                path: {
                    san: components["parameters"]["san"];
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Successful response */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["PlanResponse"];
                    };
                };
                /** @description Error Response */
                default: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["Errors"];
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/{san}/all-plans": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get All Plans
         * @description Get all plans an Agent is allowed to sell
         */
        get: {
            parameters: {
                query: {
                    fips: string;
                    state: string;
                    /** @description ISO Format Date, defaults to today's date. */
                    as_of_date: string;
                    /** @description Include careplus plans in the response */
                    careplus?: boolean;
                    /** @description Include med supp plans in the response */
                    med_supp?: boolean;
                };
                header: {
                    /** @description API Key */
                    "x-humana-api-key": components["parameters"]["api-key"];
                };
                path: {
                    san: components["parameters"]["san"];
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Successful response */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["AllPlanResponse"];
                    };
                };
                /** @description Error Response */
                default: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["Errors"];
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/{san}/certifications": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get Agent Certifications
         * @description Get list of certifications for an Agent
         */
        get: {
            parameters: {
                query?: never;
                header: {
                    /** @description API Key */
                    "x-humana-api-key": components["parameters"]["api-key"];
                };
                path: {
                    san: components["parameters"]["san"];
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Successful response */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["CertificationResponse"];
                    };
                };
                /** @description Error Response */
                default: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["Errors"];
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/{san}/licenses": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get Agent Licenses
         * @description Get list of licenses for an Agent
         */
        get: {
            parameters: {
                query?: never;
                header: {
                    /** @description API Key */
                    "x-humana-api-key": components["parameters"]["api-key"];
                };
                path: {
                    san: components["parameters"]["san"];
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Successful response */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["LicenseResponse"];
                    };
                };
                /** @description Error Response */
                default: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["Errors"];
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}